# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "8BitGo 开放平台 API",
    "version": "1.0.0",
    "description": "面向第三方应用 / 设备端的开放 API。目录、直播与联机房间公开只读；跨设备游玩上报和设备开播使用用户级令牌。鉴权用 OAuth2（client_credentials 取应用级令牌，设备码流程取用户级令牌），scope 隔离。完整字段与取值均对齐 `server/src/routes/open.js` 与 `docs/esp-open-api.md`。本文件挂在 `/.well-known/openapi.json`，供 agent / 代码生成器自发现。",
    "contact": { "name": "8BitGo", "url": "https://8bitgo.com/open" }
  },
  "servers": [
    { "url": "https://8bitgo.com/api/open/v1", "description": "生产环境" }
  ],
...
  "components": {
    "securitySchemes": {
      "openToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "开放平台令牌（RS256 签发，与站内令牌互不相认）。放在 `Authorization: Bearer <token>`。用户级 scope（library.read / library.write / live.write / saves.read）只能由设备码或授权码流程获得。"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "OAuth 风格错误码，如 invalid_token / insufficient_scope / not_found / rate_limited" },
          "error_description": { "type": "string" },
          "error_uri": { "type": "string", "format": "uri" },
          "scope": { "type": "string", "description": "部分 403 响应会回缺失的 scope" }
        },
        "required": ["error"]
      },
...
      "MeResponse": {
        "type": "object",
        "properties": {
          "client_id": { "type": "string", "description": "应用 AppID（app_ + 24 位十六进制）" },
          "kind": { "type": "string", "enum": ["app", "user"] },
          "user_id": { "type": ["string", "null"], "description": "用户级令牌才有，应用级为 null" },
          "scope": { "type": "string" },
          "expires_at": { "type": "string", "format": "date-time" }
        },
        "required": ["client_id", "kind", "scope", "expires_at"]
      },
...
      "Game": {
        "type": "object",
        "description": "对外白名单形状（open/mapper.js），不含 ROM 真实地址等内部字段",
        "properties": {
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "lang_requested": { "type": "string", "description": "请求时传的 lang" },
          "lang_actual": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "description": { "type": "string" }
            }
          },
          "platform": { "type": "string" },
          "requires_windows": { "type": "boolean", "description": "DOS 游戏是否需要 Windows 3.x / 95 / 98 客体；对应后台的 Windows 复选框。其他平台为 false" },
          "genres": { "type": "array", "items": { "type": "string" } },
          "tags": { "type": "array", "items": { "type": "string" } },
          "year": { "type": "integer" },
          "developer": { "type": "string" },
          "players": { "type": "integer" },
          "multiplayer": { "type": "boolean" },
          "icon": { "type": "string", "description": "emoji 兜底图标" },
          "cover": { "type": ["string", "null"], "format": "uri" },
          "rating": { "type": "number" },
          "rating_count": { "type": "integer" },
          "plays": { "type": "integer" },
          "added_at": { "type": ["string", "null"], "format": "date" },
          "updated_at": { "type": ["string", "null"], "format": "date-time" },
          "adult": { "type": "boolean" },
          "rom_langs": { "type": "array", "items": { "type": "string" }, "description": "可选 ROM 语言码，* 表示通用件" }
        },
        "required": ["slug", "title", "platform", "requires_windows"]
      },
      "GamePage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Game" } },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" }
        },
        "required": ["items", "page", "page_size", "total", "total_pages"]
      },
...
      "SaveMeta": {
        "type": "object",
        "properties": {
          "runtime": { "type": "string" },
          "game_slug": { "type": "string" },
          "slot": { "type": "integer" },
          "size": { "type": "integer", "description": "字节" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["runtime", "game_slug", "slot", "size"]
      },
      "CoverGame": {
        "type": "object",
        "description": "封面用的瘦身 Game（合集四宫格 / 轮播）",
        "properties": {
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "titleZh": { "type": "string" },
          "platform": { "type": "string" },
          "icon": { "type": "string" },
          "cover": { "type": ["string", "null"], "format": "uri" },
          "video": { "type": ["string", "null"] }
        },
        "required": ["slug", "title", "platform", "icon"]
      },
...
      "CollectionSummary": {
        "type": "object",
        "description": "合集元信息（不含里面的游戏，见 /v1/collections/{id}）",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "kind": { "type": "string" },
          "description": { "type": "string" },
          "gameCount": { "type": "integer" },
          "viewCount": { "type": "integer", "description": "多少人看过（去重）" },
          "covers": { "type": "array", "items": { "$ref": "#/components/schemas/CoverGame" } },
          "author": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "nickname": { "type": "string" },
              "avatar": { "type": "string" }
            }
          },
          "hidden": { "type": "boolean" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "title", "gameCount", "author", "updatedAt", "createdAt"]
      }
    }
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "tags": ["auth"],
        "summary": "自省：令牌身份 / scope / 过期",
        "security": [{ "openToken": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MeResponse" } } }
          },
          "401": { "description": "令牌无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/games": {
      "get": {
        "tags": ["games"],
        "summary": "游戏列表（分页 / 筛选）",
        "security": [{}, { "openToken": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 24 } },
          { "name": "platform", "in": "query", "schema": { "type": "string" }, "description": "机型 id，见 /v1/platforms；与 genre 同时传时取交集，再分页" },
          { "name": "requires_windows", "in": "query", "schema": { "type": "boolean" }, "description": "false 排除需要 Windows 3.x / 95 / 98 客体的 DOS 游戏；true 只返回这类游戏。不传则不筛选。筛选在分页和 total 计算前完成" },
          { "name": "genre", "in": "query", "schema": { "type": "string" }, "description": "类型 id，见 /v1/genres；与 platform 同时传时取交集，再分页" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "搜索词" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["popular", "newest", "name", "rating", "home"] } },
          { "name": "lang", "in": "query", "schema": { "type": "string" }, "description": "语言码，默认 en（站内 x-default）" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GamePage" } } }
          },
          "401": { "description": "带了 Authorization 头但令牌无效 / 已过期（不带头 = 匿名，正常返回）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "超出配额，看 Retry-After", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/saves": {
      "get": {
        "tags": ["user-data"],
        "summary": "存档清单（用户级令牌，仅元信息）",
        "security": [{ "openToken": ["saves.read"] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/SaveMeta" } } } } } }
          },
          "401": { "description": "令牌无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "需要用户级令牌", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/collections": {
      "get": {
        "tags": ["catalog"],
        "summary": "公开合集列表（分页）",
        "security": [{}, { "openToken": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 48, "default": 24 } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/CollectionSummary" } }, "page": { "type": "integer" }, "page_size": { "type": "integer" }, "total": { "type": "integer" }, "total_pages": { "type": "integer" } } } } }
          },
          "401": { "description": "令牌无效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}